wip: incorrect address

+++ src/main/java/com/wavesplatform/crypto/ChainId.java
@@ -2,9 +2,9 @@ package com.wavesplatform.crypto;
 
 public class ChainId {
 
-    public static final short MAINNET = 'W';
-    public static final short TESTNET = 'T';
-    public static final short STAGENET = 'S';
+    public static final byte MAINNET = 'W';
+    public static final byte TESTNET = 'T';
+    public static final byte STAGENET = 'S';
 
     //TODO as object?
     //TODO custom?
